Move redisdb tests to own file
This change moves the redisdb integration tests to their own file.
This should improve readability of the submodules.
